app/services/check_models.py
import os
from groq import Groq
from dotenv import load_dotenv

load_dotenv()

client = Groq(api_key=os.getenv("GROQ_API_KEY"))

models = client.models.list()

print("\nAVAILABLE MODELS:\n")
for m in models.data:
    print(m.id)
