requirement.txt
absl-py==2.3.1
acres==0.5.0
alembic==1.17.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
app_utils==0.1a4
asgiref==3.11.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
bcrypt==5.0.0
black==25.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.2.1
cmdstanpy==1.2.5
cohere==5.18.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
distro==1.9.0
Django==5.2.8
django-stubs==5.2.8
django-stubs-ext==5.2.8
ecdsa==0.19.1
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.118.0
fastavro==1.12.0
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.60.1
fsspec==2025.9.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.75.1
h11==0.16.0
h5py==3.14.0
holidays==0.81
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.1
idna==3.10
importlib_resources==6.5.2
iniconfig==2.3.0
isodate==0.7.2
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==2.15.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.77
langchain-google-genai==2.1.12
langchain-openai==0.3.34
langchain-text-splitters==0.3.11
langsmith==0.4.31
libclang==18.1.1
looseversion==1.3.0
lxml==6.0.2
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml-dtypes==0.2.0
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.5.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==1.26.4
oauthlib==3.3.1
openai==2.0.0
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
passlib==1.7.4
pathlib==1.0.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
prophet==1.1.7
proto-plus==1.26.1
protobuf==6.32.1
prov==2.1.1
psutil==7.1.0
psycopg2-binary==2.9.11
pure_eval==0.2.3
puremagic==1.30
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.4
pyparsing==3.2.5
PyPDF2==3.0.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
pywin32==311
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.1.0
rdflib==7.2.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruff==0.11.7
scipy==1.15.3
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.27
sqlparse==0.5.4
stanio==0.5.1
starlette==0.48.0
streamlit==1.49.1
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tiktoken==0.11.0
tokenizers==0.22.1
toml==0.10.2
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.3
wrapt==1.14.2
zstandard==0.25.0
